requirements.txt
confluent-kafka
python-dotenv
elasticsearch